+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  "config": path.resolve('./config/config.json'),
-  "models-path": path.resolve('./database/models'),
-  "seeders-path": path.resolve('./database/seeders'),
-  "migrations-path": path.resolve('./database/migrations')
+  "config": path.resolve('./config.json'),
+  "models-path": path.resolve('./models'),
+  "seeders-path": path.resolve('./seeders'),
+  "migrations-path": path.resolve('./migrations')
 };
