+++ package.json
@@ -1,7 +1,6 @@
 {
-  "name": "v3todo",
+  "name": "todoapp",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -9,7 +8,7 @@
     "postgres": "sudo service postgresql"
   },
   "keywords": [],
-  "author": "",
+  "author": "Dusk Kitil",
   "license": "ISC",
   "devDependencies": {
     "nodemon": "^2.0.15"
@@ -22,5 +21,17 @@
     "pg-hstore": "^2.3.4",
     "sequelize": "^6.9.0",
     "twilio": "^3.71.1"
-  }
+  },
+  "bin": {
+    "todoapp": "bin/server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/duskkitil/todoApp.git"
+  },
+  "bugs": {
+    "url": "https://github.com/duskkitil/todoApp/issues"
+  },
+  "homepage": "https://github.com/duskkitil/todoApp#readme",
+  "description": ""
 }
